hw/riscv: Fix the error of parameter type mismatch

+++ hw/riscv/nuclei_demosoc.c
@@ -342,7 +342,7 @@ static void create_fdt(DemoSoCState *s, const struct MemmapEntry *memmap,
     g_free(nodename);
 
     nodename = g_strdup_printf("/soc/spi@%lx",
-                               memmap[DEMOSOC_QSPI0].base);
+                               (long)memmap[DEMOSOC_QSPI0].base);
     qemu_fdt_add_subnode(fdt, nodename);
     qemu_fdt_setprop_string(fdt, nodename, "compatible", "nuclei,spi0");
     qemu_fdt_setprop_cells(fdt, nodename, "reg",
