+++ hw/riscv/nuclei_evalsoc.c
@@ -361,7 +361,7 @@ static void create_fdt(EvalSoCState *s, const struct MemmapEntry *memmap,
     g_free(nodename);
 
     nodename = g_strdup_printf("/soc/spi@%lx",
-                               memmap[EVALSOC_QSPI0].base);
+                               (long)memmap[EVALSOC_QSPI0].base);
     qemu_fdt_add_subnode(fdt, nodename);
     qemu_fdt_setprop_string(fdt, nodename, "compatible", "nuclei,spi0");
     qemu_fdt_setprop_cells(fdt, nodename, "reg",
